+++ dashboard/src/actions/__tests__/toastActions.test.js
@@ -1,35 +1,6 @@
 import * as ACTIONS from "../toastActions";
 
-// import configureStore to create a mock store where we will dispatch our actions
-import configureStore from "redux-mock-store";
-// import thunk middle to make our action asynchronous
-import thunk from "redux-thunk";
-
-// initialise middlewares
-const middlewares = [thunk];
-// initialise MockStore which is only the configureStore method which take middlewares as its parameters
-
-const mockStore = configureStore(middlewares);
-
-const mockState = {
-  toastReducer: {
-    alerts: [
-      {
-        variant: "success",
-        title: "Logged in successfully!",
-        key: "test1",
-        message: "",
-      },
-      {
-        variant: "success",
-        title: "Saved!",
-        key: "test2",
-        message: "",
-      },
-    ],
-  },
-};
-const store = mockStore(mockState);
+import { mockState, store } from "store/mockStore";
 
 describe("Toast Actions", () => {
   beforeEach(() => {
@@ -43,35 +14,20 @@ describe("Toast Actions", () => {
     jest.resetAllMocks();
   });
   test("Show Failure Toast", () => {
-    const expectedActions = [
-      {
-        type: "SHOW_TOAST",
-        payload: [
-          {
-            variant: "success",
-            title: "Logged in successfully!",
-            key: "test1",
-            message: "",
-          },
-          {
-            variant: "success",
-            title: "Saved!",
-            key: "test2",
-            message: "",
-          },
-          {
-            key: expect.anything(),
-            variant: "danger",
-            message: "Please try again later",
-            title: "Something went wrong",
-          },
-        ],
-      },
-    ];
+    const expectedOutput = {
+      key: expect.anything(),
+      variant: "danger",
+      message: "Please try again later",
+      title: "Something went wrong",
+    };
+    store.dispatch(ACTIONS.showFailureToast());
+    const bestMatch = (output) => {
+      return output.find((item) => item.title.includes("Something went wrong"));
+    };
 
-    return store.dispatch(ACTIONS.showFailureToast()).then(() => {
-      expect(store.getActions()).toEqual(expectedActions);
-    });
+    const givenPayload = store.getActions()[0].payload;
+    const actions = bestMatch(givenPayload);
+    expect(actions).toEqual(expectedOutput);
   });
 
   test("Show Toast", () => {
@@ -83,12 +39,12 @@ describe("Toast Actions", () => {
     };
 
     store.dispatch(ACTIONS.showToast("success", "Data updated!"));
-    const bestMacth = (output) => {
+    const bestMatch = (output) => {
       return output.find((item) => item.title === "Data updated!");
     };
 
     const givenPayload = store.getActions()[0].payload;
-    const actions = bestMacth(givenPayload);
+    const actions = bestMatch(givenPayload);
     expect(actions).toEqual(expectedOutput);
   });
   test("Show Session Expired", () => {
@@ -107,21 +63,22 @@ describe("Toast Actions", () => {
     const actions = bestMatch(givenPayload);
     expect(actions).toEqual(expectedOutput);
   });
-  test("Hide toast", async () => {
-    await store.dispatch(ACTIONS.hideToast("test2"));
-    const expectedActions = [
+  test("Hide toast", () => {
+    const keyToHide = "test2";
+    const expected = [
       {
-        type: "SHOW_TOAST",
-        payload: [
-          {
-            variant: "success",
-            title: "Logged in successfully!",
-            key: "test1",
-            message: "",
-          },
-        ],
+        variant: "success",
+        title: "Saved!",
+        key: "test2",
+        message: "",
       },
     ];
-    expect(store.getActions()).toEqual(expectedActions);
+    store.dispatch(ACTIONS.hideToast(keyToHide));
+    const bestMatch = (output) => {
+      return output.filter((item) => item.filter !== keyToHide);
+    };
+    const givenPayload = store.getActions()[0].payload;
+    const actions = bestMatch(givenPayload);
+    expect(actions).toEqual(expect.not.arrayContaining(expected));
   });
 });
