+++ dashboard/src/modules/components/ToastNotificationComponent/index.test.jsx
@@ -1,27 +1,10 @@
 import * as reactRedux from "react-redux";
 
+import { mockState, store } from "store/mockStore";
+
 import React from "react";
 import ToastComponent from "./index";
-import configureStore from "redux-mock-store";
 import { render } from "@testing-library/react";
-
-// test block
-const mockStore = configureStore();
-const initialState = {
-  alerts: [
-    {
-      variant: "success",
-      title: "Logged in successfully!",
-      key: 112,
-    },
-    {
-      variant: "success",
-      title: "Saved!",
-      key: 114,
-    },
-  ],
-};
-const store = mockStore(initialState);
 
 describe("Test Toast Component", () => {
   beforeEach(() => {
@@ -29,14 +12,14 @@ describe("Test Toast Component", () => {
     const useSelectorMock = jest.spyOn(reactRedux, "useSelector");
     const useDispatchMock = jest.spyOn(reactRedux, "useDispatch");
 
-    useDispatchMock.mockReturnValue(mockDispatchFn);
+    useDispatchMock.mockReturnValueOnce(mockDispatchFn);
 
-    useSelectorMock.mockReturnValue(initialState.alerts);
+    useSelectorMock.mockReturnValueOnce(mockState.toastReducer);
   });
   // render the component on virtual dom
   it("Toast", () => {
     render(<ToastComponent store={store} />, {
-      initialState,
+      mockState,
     });
 
     expect(store.getActions()).toMatchSnapshot();
